src/requirements.txt
kafka-python
smhasher
bitarray

